Add migration to include Stock column in Vehicle table and change BatteryCapacity type to int
- Created migration 20251017145246_AddStockColumnToVehicleTable to add Stock column to Vehicles.
- Created migration 20251017154735_ChangeBatteryCapacityToInt to change BatteryCapacity from decimal to int.
- Created migration 20251017155318_int-BatteryCapacity for further adjustments to BatteryCapacity.

+++ EVDealerSales.BusinessObject/DTOs/VehicleDTOs/VehicleFilterDto.cs
@@ -6,8 +6,8 @@ namespace EVDealerSales.BusinessObject.DTOs.VehicleDTOs
         public string? SearchTerm { get; set; }
 
         // Filter by Battery Capacity
-        public decimal? MinBatteryCapacity { get; set; }
-        public decimal? MaxBatteryCapacity { get; set; }
+        public int? MinBatteryCapacity { get; set; }
+        public int? MaxBatteryCapacity { get; set; }
 
         // Filter by Top Speed
         public int? MinTopSpeed { get; set; }
